+++ src/data/providers/balancer-api/index.ts
@@ -5,7 +5,10 @@ import { NestedPools } from './modules/nested-pool-state';
 import { SorSwapPaths } from './modules/sorSwapPaths';
 
 export { SorInput as GetQuoteInput } from './modules/sorSwapPaths';
-export { mapPoolToNestedPoolState } from './modules/nested-pool-state';
+export {
+    mapPoolToNestedPoolState,
+    PoolGetPool,
+} from './modules/nested-pool-state';
 
 export class BalancerApi {
     balancerApiClient: BalancerApiClient;