+++ test/anvil/anvil-global-setup.ts
@@ -21,6 +21,8 @@ type NetworksWithFork = Extract<
     | 'OPTIMISM'
     | 'MODE'
     | 'FRAXTAL'
+    | 'AVALANCHE'
+    | 'ARBITRUM_ONE'
 >;
 
 const ANVIL_PORTS: Record<NetworksWithFork, number> = {
@@ -32,6 +34,8 @@ const ANVIL_PORTS: Record<NetworksWithFork, number> = {
     OPTIMISM: 9045,
     MODE: 9145,
     FRAXTAL: 9245,
+    AVALANCHE: 9345,
+    ARBITRUM_ONE: 9445,
 };
 
 export const ANVIL_NETWORKS: Record<NetworksWithFork, NetworkSetup> = {
@@ -79,6 +83,18 @@ export const ANVIL_NETWORKS: Record<NetworksWithFork, NetworkSetup> = {
         fallBackRpc: 'https://fraxtal.gateway.tenderly.co/',
         port: ANVIL_PORTS.FRAXTAL,
         forkBlockNumber: 7164945n,
+    },
+    AVALANCHE: {
+        rpcEnv: 'AVALANCHE_RPC_URL',
+        fallBackRpc: 'https://avalanche.gateway.tenderly.co/',
+        port: ANVIL_PORTS.AVALANCHE,
+        forkBlockNumber: 48164407n,
+    },
+    ARBITRUM_ONE: {
+        rpcEnv: 'ARBITRUM_ONE_RPC_URL',
+        fallBackRpc: 'https://arbitrum.gateway.tenderly.co/',
+        port: ANVIL_PORTS.ARBITRUM_ONE,
+        forkBlockNumber: 234936318n,
     },
 };